# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Motorcycle ERP & POS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parts:
    get:
      operationId: listParts
      tags: [parts]
      summary: List parts inventory
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search by SKU, name, or model compatibility
        - in: query
          name: warehouseId
          schema:
            type: integer
        - in: query
          name: condition
          schema:
            $ref: "#/components/schemas/PartCondition"
        - in: query
          name: lowStock
          schema:
            type: boolean
          description: Filter parts with quantity <= reorder point
      responses:
        "200":
          description: List of parts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Part"
    post:
      operationId: createPart
      tags: [parts]
      summary: Create a part
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartBody"
      responses:
        "201":
          description: Part created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"
  /parts/{id}:
    get:
      operationId: getPart
      tags: [parts]
      summary: Get part by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"
    put:
      operationId: updatePart
      tags: [parts]
      summary: Update part
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePartBody"
      responses:
        "200":
          description: Updated part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"
    delete:
      operationId: deletePart
      tags: [parts]
      summary: Delete part
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── MOTORCYCLES ───────────────────────────────────────────────────────────
  /motorcycles:
    get:
      operationId: listMotorcycles
      tags: [motorcycles]
      summary: List motorcycle inventory
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/MotorcycleStatus"
        - in: query
          name: warehouseId
          schema:
            type: integer
      responses:
        "200":
          description: List of motorcycles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Motorcycle"
    post:
      operationId: createMotorcycle
      tags: [motorcycles]
      summary: Create a motorcycle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMotorcycleBody"
      responses:
        "201":
          description: Motorcycle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Motorcycle"
  /motorcycles/{id}:
    get:
      operationId: getMotorcycle
      tags: [motorcycles]
      summary: Get motorcycle by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Motorcycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Motorcycle"
    put:
      operationId: updateMotorcycle
      tags: [motorcycles]
      summary: Update motorcycle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMotorcycleBody"
      responses:
        "200":
          description: Updated motorcycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Motorcycle"
    delete:
      operationId: deleteMotorcycle
      tags: [motorcycles]
      summary: Delete motorcycle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── VENDORS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ─── ENUMS ───────────────────────────────────────────────────────────────
...
    PartCondition:
      type: string
      enum: [new, used]
    MotorcycleStatus:
      type: string
      enum: [available, sold, in_service, pre_owned]
...
    Part:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        condition:
          $ref: "#/components/schemas/PartCondition"
        modelCompatibility:
          type: string
        quantityOnHand:
          type: integer
        reorderPoint:
          type: integer
        costPrice:
          type: number
        sellingPrice:
          type: number
        warehouseId:
          type: integer
        binId:
          type: integer
        binLabel:
          type: string
        warehouseName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sku
        - name
        - condition
        - quantityOnHand
        - reorderPoint
        - costPrice
        - sellingPrice
        - createdAt
    CreatePartBody:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        condition:
          $ref: "#/components/schemas/PartCondition"
        modelCompatibility:
          type: string
        quantityOnHand:
          type: integer
        reorderPoint:
          type: integer
        costPrice:
          type: number
        sellingPrice:
          type: number
        warehouseId:
          type: integer
        binId:
          type: integer
      required:
        - sku
        - name
        - condition
        - quantityOnHand
        - reorderPoint
        - costPrice
        - sellingPrice
    UpdatePartBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        condition:
          $ref: "#/components/schemas/PartCondition"
        modelCompatibility:
          type: string
        quantityOnHand:
          type: integer
        reorderPoint:
          type: integer
        costPrice:
          type: number
        sellingPrice:
          type: number
        warehouseId:
          type: integer
        binId:
          type: integer

    # ─── MOTORCYCLE ──────────────────────────────────────────────────────────
    Motorcycle:
      type: object
      properties:
        id:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        vin:
          type: string
        color:
          type: string
        engineSize:
          type: string
        mileage:
          type: integer
        status:
          $ref: "#/components/schemas/MotorcycleStatus"
        costPrice:
          type: number
        sellingPrice:
          type: number
        warehouseId:
          type: integer
        warehouseName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - make
        - model
        - year
        - status
        - costPrice
        - sellingPrice
        - createdAt
    CreateMotorcycleBody:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        vin:
          type: string
        color:
          type: string
        engineSize:
          type: string
        mileage:
          type: integer
        status:
          $ref: "#/components/schemas/MotorcycleStatus"
        costPrice:
          type: number
        sellingPrice:
          type: number
        warehouseId:
          type: integer
      required:
        - make
        - model
        - year
        - status
        - costPrice
        - sellingPrice
    UpdateMotorcycleBody:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        vin:
          type: string
        color:
          type: string
        engineSize:
          type: string
        mileage:
          type: integer
        status:
          $ref: "#/components/schemas/MotorcycleStatus"
        costPrice:
          type: number
        sellingPrice:
          type: number
        warehouseId:
          type: integer

    # ─── VENDOR ──────────────────────────────────────────────────────────────
...